Move includeFile to ComponentPreprocessor
The entry point to ComponentPreprocessor resides in the includeFile
method, which resides in Parser.

Let's move this to the ComponentPreprocessor class to benefit from the
instance properties, and also to consolidate include functionality.
